Separated pure image compression from checking if it should happen at all
