+++ collect_app/src/androidTest/java/org/odk/collect/android/instrumented/utilities/ImageCompressorTest.kt
