+++ collect_app/src/test/java/org/odk/collect/android/utilities/ImageCompressionControllerTest.kt
@@ -1,0 +1,5 @@
+package org.odk.collect.android.utilities
+
+import org.junit.jupiter.api.Assertions.*
+
+internal class ImageCompressionControllerTest
